refactored progress-bar
