+++ src/components/progress-bar/progress-bar.component.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 // UI Core
 import { makeStyles } from "@material-ui/core";
-import Typography from "@material-ui/core/Typography";
 // Components
 import CustomIcon from "../custom-icon/custom-icon.component.jsx";
+import CustomOverlayDisplay from "../custom-overlay-display/custom-overlay-display.component.jsx";
+import CustomProgress from "../custom-progress/custom-progress.component.jsx";
 
-// TODO: Rethink whether to refactor further
 const ProgressBar = ({ variant, stat, size }) => {
-    const useStyles = makeStyles(theme => ({
+    const useStyles = makeStyles({
         bar: {
             width: "100%",
             height: "60px",
@@ -16,38 +16,14 @@ const ProgressBar = ({ variant, stat, size }) => {
             backgroundColor: "transparent",
             position: "relative",
         },
-        progress: {
-            borderRadius: "20px",
-            height: "100%",
-            backgroundPosition: "center",
-            backgroundImage: `linear-gradient(${theme.palette.primary.main}, ${theme.palette.secondary.main})`,
-            transition: "width 1.2s",
-            transitionTimingFunction: "linear",
-            width: `${stat}%`,
-        },
-        statDisplay: {
-            width: "100%",
-            textAlign: "center",
-            position: "absolute",
-            top: "12px",
-            bottom: "12px",
-        },
-    }));
+    });
     const classes = useStyles();
 
     return (
-        <div className="container">
-            <div className={classes.bar}>
-                <CustomIcon size={size} variant={variant} />
-                <div className={classes.progress} />
-                <Typography
-                    variant="h4"
-                    component="h4"
-                    className={classes.statDisplay}
-                >
-                    {stat}
-                </Typography>
-            </div>
+        <div className={classes.bar}>
+            <CustomIcon size={size} variant={variant} />
+            <CustomProgress stat={stat} />
+            <CustomOverlayDisplay stat={stat} />
         </div>
     );
 };
